This release (2018.06.04) is based on the same ipc code present in kdb+3.6.
Important - these notes and libraries refer to code which is NOT to be loaded into kdb+ as a shared library.

Beginning with this release, the release date may be inquired from the library via the call
extern I ver();
which will return an int as yyyymmdd. Additionally, there are embedded strings which can be grepped out:

$strings e.so |grep "e.so"

The new features in this release are
- uds (unix domain sockets)
- ssl/tls support
- optionally allow ipc msgs larger than 2GB. The default limit remains 2GB.

To minimize dependencies for existing projects, there are now 2 sets of files available:
1) No ssl/tls capabilities nor depedencies: c.dll,c.lib,cst.dll,cst.lib,c_static.lib,cst_static.lib,c.o
2) ssl/tls capability with openssl depedency: e.dll,e.lib,est.dll,est.lib,e_static.lib,est_static.lib,e.o

The "e" set of files, those with ssl/tls support, contain all the functionality of the "c" files - do NOT link with BOTH "c" and "e" files; just choose one of them.

For windows builds: c.lib is a stub library which loads c.dll and resolves the functions dynamically. e.lib does the same for e.dll. We no longer ship c.obj or cst.obj; they have been replaced by c_static.lib and cst_static.lib, and are complemented by e_static.lib and est_static.lib - these static libraries have no dependency on the aforementioned dlls.

cst continues to represent "single-threaded" apps, those which on windows have issues due to the LoadLibrary api. see https://code.kx.com/q/interfaces/c-client-for-q/#windows-and-the-loadlibrary-api for more info.

uds
Unix domain socket may be requested via the ip address 0.0.0.0, e.g.
 int handle=khpu("0.0.0.0",5000,"user:password");

ssl/tls
To use this feature, you must link with one of the "e" libs mentioned above.
Encrypted connections may be requested via the capability parameter of the new khpunc function, e.g.
 extern I khpunc(S hostname,I port,S usernamepassword,I timeout,I capability); / capability is a bit field (1 - 1TB limit, 2 - use TLS).
 int handle=khpunc("remote host",5000,"user:password",timeout,2);
There's an additional return value for tls connections, -3, which indicates the openssl init failed. This can be checked via
 extern K sslInfo(K x); // returns an error if init fails, or a dict of settings similar to -26!x
 if(handle==-3){K x=ee(sslInfo((K)0));printf("Init error %s\n",xt==-128?x->s:"unknown");r0(x);}
A restriction for ssl/tls connections - these can be used from the initialization thread only, i.e. the thread which first calls any khp function since the start of the application.
The lib is sensitive to the same environment variables as kdb+, noted at https://code.kx.com/q/cookbook/ssl
The openssl libs are loaded dynamically,  the first time a tls connection is requested. It may be forced on startup with
 int h=khpunc("",-1,"",0,2); // remember to test the return value for -3

ipc msgs larger than 2GB
During the initial handshake of a connection, each side's capability is exchanged, and the common maximum is chosen for the connection. By setting the capability parameter for khpunc, the default msg size limit for this connection can be raised from 2GB to 1TB. e.g.
 int handle=khpunc("remote host",5000,"user:password",timeout,1);
A tls enabled connection supporting upto 1TB messages can be achieved via bit-or of the tls and 1TB bits, e.g.
 int handle=khpunc("remote host",5000,"user:password",timeout,1|2);